Move whoami command to new file and reformat output
