+++ src/utils.ts
@@ -1,0 +1,13 @@
+import { promisify } from 'util';
+
+const getCurrentProfile = () => {
+  return process.env.AWS_PROFILE || '';
+};
+
+const timeout = async (): Promise<void> => {
+  const delay = promisify(setTimeout);
+
+  await delay(1500);
+};
+
+export { getCurrentProfile, timeout };
